third_party/WebKit/LayoutTests/crypto/worker-random-values-limits-expected.txt

[Worker] Tests the limits of crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/random-values-limits.js
PASS [Worker] 'crypto' in self is true
PASS [Worker] 'getRandomValues' in self.crypto is true
PASS [Worker] crypto.getRandomValues(almostTooLargeArray) did not throw exception.
PASS [Worker] crypto.getRandomValues(tooLargeArray) threw exception QuotaExceededError: An attempt was made to add something to storage that exceeded the quota..
PASS successfullyParsed is true

TEST COMPLETE
